Initial server for sathunt
This is the starting libraries for the sathunt server in a broken state.
The database and iod are included until there is a fairly static module
on github to reference. This is also for testing purposes until the
individual endpoints are broken out into their own files which will
require those broken out iod and database modules.